requirements.txt
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
greenlet==3.2.4
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.47
starlette==0.49.3
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
